working version

+++ UsbSerial/lib/Settings/Settings.h
@@ -2,15 +2,15 @@
 #define SETTINGS_H
 
 struct Settings {
-    double frontMax = 5.00;
-    double backMax = 4.60;
-    double rideFront = 3.00;
-    double rideBack = 2.50;
-    double frontUpX = 1.00;
-    double frontDownX = 1.66;
-    double backUpX = 0.50;
-    double backDownX = 1.66;
-    double parkDuration = 4.0;
+    float frontMax = 5.00;
+    float backMax = 4.60;
+    float rideFront = 3.00;
+    float rideBack = 2.50;
+    float frontUpX = 1.00;
+    float frontDownX = 1.66;
+    float backUpX = 0.50;
+    float backDownX = 1.66;
+    float parkDuration = 4.0;
 };
 
 #endif //SETTINGS_H
